location verification change
